refactor: modify package structure(apply MVC Pattern):
