+++ src/main/java/racingcar/controller/MainController.java
@@ -1,4 +1,6 @@
-package racingcar;
+package racingcar.controller;
+
+import racingcar.domain.Game;
 
 public class MainController {
