+++ src/main/java/racingcar/util/InputValidator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 import java.util.Arrays;
 import java.util.HashSet;
