+++ src/main/java/racingcar/view/Output.java
@@ -1,6 +1,7 @@
-package racingcar;
+package racingcar.view;
 
 import java.util.List;
+import racingcar.domain.Car;
 
 public class Output {
