+++ src/test/java/racingcar/InputValidatorTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.util.InputValidator;
 
 public class InputValidatorTest {
